+++ FunctionsHW8.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <iostream>
+
+void printNaturalNumbersA(int count);
+void printNaturalNumbersB(int count);
+void printNaturalNumbersC(int count);
+void printNaturalNumbersD(int count);
+void helpFunctionD(int count, int);
+
+int factorialRec(unsigned int count);
+int factorialLoop(unsigned int count);
+
+void countBySign(int n);
